+++ lib/contest.c
@@ -1,7 +1,7 @@
 /**
  * @file contest.c
  * @author Matthew Getgen
- * @brief Logic for running a game.
+ * @brief Logic for running a contest.
  * @date 2026-05-12
  */
