+++ lib/platforms/platform.h
@@ -10,28 +10,10 @@
 
 #include "../battleshipslib.h"
 
-// NOTE(mattg): Also be sure to implement BShip_Arena_Initialize and BShip_Arena_Destroy.
-// They are exposed via the library header, which is why they aren't here.
 
-#define BSHIP_ARENA_PUSH(arena, type) \
-    ((type *)BShip_Arena_Push(arena, sizeof(type)))
+void *BShip_Allocate(size_t size);
 
-#define BSHIP_ARENA_PUSH_ARRAY(arena, type, count) \
-    ((type *)BShip_Arena_Push(arena, sizeof(type) * (count)))
-
-#define BSHIP_ARENA_TEMP_BEGIN(arena) \
-    BShip_ArenaMark __mark = BShip_ArenaMark_Get(arena)
-
-#define BSHIP_ARENA_TEMP_END(arena) \
-    BShip_Arena_Rollback(arena, __mark)
-
-void *BShip_Arena_Push(BShip_Arena *arena, size_t size);
-
-void BShip_Arena_Reset(BShip_Arena *arena);
-
-BShip_ArenaMark BShip_ArenaMark_Get(BShip_Arena *arena);
-
-void BShip_Arena_Rollback(BShip_Arena *arena, BShip_ArenaMark mark);
+void BShip_Deallocate(void *ptr);
 
 
 typedef struct BShip_Connection BShip_Connection;
